remote_lollms/requirements.txt
python-socketio
requests
websocket-client